ServerlessでAWS Lambdaにデプロイできるようにした(#1 予約投稿のキーワードフォーマットを環境変数から読み込むようにした(#2

+++ src/main.py
@@ -5,14 +5,15 @@ from datetime import datetime as dt
 import os
 
 
-class ReservationPostQiita(object):
+class ReservationPostQiitaClient(object):
 
   def __init__(self, access_token):
     self.access_token = access_token
     self.resrvation_items = []
-    self.client = QiitaClient(access_token=os.environ['QIITA_ACCESS_TOKEN'])
+    self.client = QiitaClient(access_token=access_token)
     self._get_all_items()
     print(f'all_items: {len(self.items)}')
+
 
   def _get_all_items(self):
     self.items = []
@@ -28,7 +29,9 @@ class ReservationPostQiita(object):
         break
 
 
-  def get_reservation_items(self, post_datetime):
+  def get_reservation_items(self, reservation_keyword):
+
+    self.resrvation_items = []
 
     # 限定共有投稿に絞り込む
     private_items = list(filter(lambda x: x['private'] == True, self.items))
@@ -39,11 +42,9 @@ class ReservationPostQiita(object):
     print(f'commented items: {len(comments_items)}')
 
     # 記事コメントに予約投稿キーワードが含まれる記事に絞り込む
-    self.resrvation_items = []
     for item in comments_items:
       comments = self.client.list_item_comments(item['id']).to_json()
-      keyword = f'予約投稿:{post_datetime}'
-      resrvation_comments = list(filter(lambda x: keyword in x['body'], comments))
+      resrvation_comments = list(filter(lambda x: reservation_keyword in x['body'], comments))
       if len(resrvation_comments) > 0:
         self.resrvation_items.append(item)
 
@@ -52,32 +53,25 @@ class ReservationPostQiita(object):
 
   def post_items(self):
     for item in self.resrvation_items:
-      # 記事を投稿する
+      # 記事を公開投稿する
       postParams = {}
       for key in ['title', 'body', 'tags']:
         postParams[key] = item[key]
-      postParams['title'] = '【自動投稿】' + postParams['title']
-      postParams['private'] = True # Test
+      postParams['title'] = '【予約投稿】' + postParams['title']
+      postParams['private'] = True # 本利用する際にはFalseにする
       post_res = self.client.create_item(postParams)
-      print(post_res.status)
+
+      if post_res.status != 201:
+        print(f'error status: {post_res.status}')
+        # エラーの場合はとりあえず落とす
+        return
       print('posted: ' + item['title'])
 
       # 限定共有投稿を削除する
       del_res = self.client.delete_item(item['id'])
-      print(del_res.status)
+      if del_res.status != 204:
+        print(f'error status: {post_res.status}')
+        return
       print('deleted: '  + item['title'])
 
       # TODO: twitterに投稿する
-
-
-if __name__ == '__main__':
-  client = ReservationPostQiita('')
-
-  # 予約投稿を取得する
-  datetime = dt.now().strftime('%Y/%m/%d %H:00')
-  client.get_reservation_items(datetime)
-
-  # 記事を投稿する
-  client.post_items()
-
-  # TODO: slackに通知する
